imapclient: implement Client.Closed
When using a context to close connections, the goroutine blocking on the
context to close the connection will keep a reference to the connection,
so it is never free'd from memory. For long running processes, this
means that we accumulate connections indefinitely.

Implement a Client.Closed() to be used by clients to monitor when a
connection is closed. This likely has other use-cases.